Enable unconvert linter (#3171)
* Enable unconvert linter

This linter's doc describes it as:

   The unconvert program analyzes Go packages to identify unnecessary
   type conversions; i.e., expressions T(x) where x already has type T.


* Unrestrict the number of linter warnings

It is best to show all warnings at once than to reveal them piece-meal,
particularly in CI where the feedback loop can be a bit slow.

By default, linters may only print the same message three times
(https://golangci-lint.run/usage/configuration/#issues-configuration)
The unconvert linter always prints the same message, so it specially
affected by this setting.


* Remove redundant type conversions


---------

+++ pkg/controller/datavolume/import-controller_test.go
@@ -1700,7 +1700,7 @@ var _ = Describe("All DataVolume Tests", func() {
 		largeOverhead   = float64(0.75)
 	)
 	DescribeTable("GetRequiredSpace should return properly enlarged sizes,", func(imageSize int64, overhead float64) {
-		for testedSize := int64(imageSize - 1024); testedSize < imageSize+1024; testedSize++ {
+		for testedSize := imageSize - 1024; testedSize < imageSize+1024; testedSize++ {
 			alignedImageSpace := imageSize
 			if testedSize > imageSize {
 				alignedImageSpace = imageSize + Mi
